Makes the typeahead functionality a bit more reusable.
